+++ tests/fixtures_utils.py
@@ -1,20 +1,25 @@
 # -*- coding: utf-8 -*-
 import csv
-import inspect
 import os
 from typing import List
 from typing import Tuple
 
 from mantarray_waveform_analysis import peak_detection
+from mantarray_waveform_analysis import peak_detector
+from mantarray_waveform_analysis import PipelineTemplate
+import matplotlib
+import matplotlib.pyplot as plt
 import numpy as np
 import pytest
+from stdlib_utils import get_current_file_abs_directory
 
-PATH_OF_CURRENT_FILE = os.path.dirname((inspect.stack()[0][1]))
+
+matplotlib.use("Agg")
+PATH_OF_CURRENT_FILE = get_current_file_abs_directory()
+
 
 PATH_TO_DATASETS = os.path.join(PATH_OF_CURRENT_FILE, "datasets")
 PATH_TO_PNGS = os.path.join(PATH_OF_CURRENT_FILE, "pngs")
-
-CONVERSION_TO_INT = 100000
 
 
 def _load_file(file_path: str) -> Tuple[List[str], List[str]]:
@@ -55,28 +60,58 @@ def create_numpy_array_of_raw_gmr_from_python_arrays(time_array, gmr_array):
 
 
 def _run_peak_detection(filename, sampling_rate_construct=100, flip_data=True):
-    sampling_rate = sampling_rate_construct
-    time_series = np.arange(
-        0, 1000000, int(CONVERSION_TO_INT / sampling_rate_construct)
-    )
-    my_local_path_data_2 = os.path.join(PATH_TO_DATASETS, filename)
-
-    time, v = _load_file_tsv(my_local_path_data_2)
+    time, v = _load_file_tsv(os.path.join(PATH_TO_DATASETS, filename))
 
     # create numpy matrix
-    data = create_numpy_array_of_raw_gmr_from_python_arrays(time, v)
+    raw_data = create_numpy_array_of_raw_gmr_from_python_arrays(time, v)
+    simple_pipeline_template = PipelineTemplate(
+        tissue_sampling_period=1 / sampling_rate_construct * 100000
+    )
+    pipeline = simple_pipeline_template.create_pipeline()
+    pipeline.load_raw_gmr_data(raw_data, raw_data)
+    filtered_data = pipeline.get_noise_filtered_gmr()
+    peak_and_valley_timepoints = peak_detector(
+        filtered_data, twitches_point_up=not flip_data
+    )
+    return pipeline, peak_and_valley_timepoints
 
-    # flip the data
-    if flip_data:
-        data = peak_detection.convert_voltage_to_displacement(data)
 
-    # noise cancellation
-    noise_free_data, sampling_rate = peak_detection.noise_filtering(data, sampling_rate)
+def _plot_data(peak_and_valley_indices, filtered_data, my_local_path_graphs):
+    time_series = filtered_data[0, :]
+    peak_indices, valley_indices = peak_and_valley_indices
+    plt.figure()
+    plt.plot(time_series, filtered_data[1, :], "b")
+    plt.plot(
+        time_series[peak_indices],
+        filtered_data[1][peak_indices],
+        "ro",
+        label="peaks",
+        fillstyle="none",
+    )
+    plt.plot(
+        time_series[valley_indices],
+        filtered_data[1][valley_indices],
+        "go",
+        label="valleys",
+        fillstyle="none",
+    )
+    plt.xlabel("Time (centimilliseconds)")
+    plt.ylabel("Magnetic Signal")
+    plt.legend(bbox_to_anchor=(0, -0.25), loc="lower center")
+    plt.tight_layout()
+    plt.savefig(my_local_path_graphs)
+    plt.close()
 
-    # call peak_detection
-    peakind = peak_detection.peak_detector(noise_free_data, sampling_rate, data)
 
-    return data, time_series, peakind, sampling_rate, noise_free_data
+def _get_data_metrics(well_fixture):
+    pipeline, peak_and_valley_indices = well_fixture
+    filtered_data = pipeline.get_noise_filtered_gmr()
+    return peak_detection.data_metrics(peak_and_valley_indices, filtered_data)
+
+
+def assert_percent_diff(actual, expected, threshold=0.0006):
+    percent_diff = abs(actual - expected) / expected
+    assert percent_diff < threshold
 
 
 @pytest.fixture(scope="session", name="raw_generic_well_a1")