+++ CHANGELOG.rst
@@ -1,5 +1,13 @@
 Changelog for Mantarray Waveform Analysis
-===================================
+=========================================
+
+0.4.0 (unreleased)
+------------------
+
+- Refactored twitch width analysis so that it interpolates to find a point to use
+- Added aggregate statistic metrics for twitch widths
+- Refactored peak detection to be more robust
+- Cached the filter coefficients in PipelineTemplate to improve performance
 
 0.3.1 (2020-08-31)
 ------------------